refactor database patching to it's own file
